+++ src/NonEmptyArray.ts
@@ -349,14 +349,14 @@ export const unzip: <A, B>(as: NonEmptyArray<[A, B]>) => [NonEmptyArray<A>, NonE
  * Prepend an element to every member of an array
  *
  * @example
- * import { cons, prependToAll } from 'fp-ts/NonEmptyArray'
+ * import { cons, prependAll } from 'fp-ts/NonEmptyArray'
  *
- * assert.deepStrictEqual(prependToAll(9)(cons(1, [2, 3, 4])), cons(9, [1, 9, 2, 9, 3, 9, 4]))
+ * assert.deepStrictEqual(prependAll(9)(cons(1, [2, 3, 4])), cons(9, [1, 9, 2, 9, 3, 9, 4]))
  *
  * @category combinators
- * @since 2.9.0
+ * @since 2.10.0
  */
-export const prependToAll: <A>(e: A) => (xs: NonEmptyArray<A>) => NonEmptyArray<A> = RNEA.prependToAll as any
+export const prependAll: <A>(e: A) => (xs: NonEmptyArray<A>) => NonEmptyArray<A> = RNEA.prependAll as any
 
 /**
  * Places an element in between members of an array
@@ -792,6 +792,15 @@ export const apS: <A, N extends string, B>(
 // -------------------------------------------------------------------------------------
 // deprecated
 // -------------------------------------------------------------------------------------
+
+/**
+ * Use `prependAll` instead.
+ *
+ * @category combinators
+ * @since 2.9.0
+ * @deprecated
+ */
+export const prependToAll = prependAll
 
 /**
  * Use `concatAll` instead.